Use ExprGen to test ExprParser
